Reduce tool redundancy for [Judge0 CE] (#3837)
