+++ api/core/tools/provider/builtin/judge0ce/judge0ce.py
@@ -1,14 +1,14 @@
 from typing import Any
 
 from core.tools.errors import ToolProviderCredentialValidationError
-from core.tools.provider.builtin.judge0ce.tools.submitCodeExecutionTask import SubmitCodeExecutionTaskTool
+from core.tools.provider.builtin.judge0ce.tools.executeCode import ExecuteCodeTool
 from core.tools.provider.builtin_tool_provider import BuiltinToolProviderController
 
 
 class Judge0CEProvider(BuiltinToolProviderController):
     def _validate_credentials(self, credentials: dict[str, Any]) -> None:
         try:
-            SubmitCodeExecutionTaskTool().fork_tool_runtime(
+            ExecuteCodeTool().fork_tool_runtime(
                 meta={
                     "credentials": credentials,
                 }
